+++ AdventOfCode.Year2019/Implementations/ShipComputer/ShipComputerFunctions/ComputerFunction_IntCode_Multiply.cs
@@ -4,7 +4,7 @@
     {
         protected override ShipComputerFunctionModel DoWork()
         {
-            Data[Data[Position + 3]] = Value2 * Value1;
+            Data[(int)Data[(int)Position + 3]] = Value2 * Value1;
 
             return new ShipComputerFunctionModel()
             {
